factor out package.json utils into a package
