+++ packages/package-json-utils/README.md
@@ -1,0 +1,3 @@
+# @auto-it/package-json-utils
+
+Shared utilities for parsing information from a package.json
