+++ plugins/npm/package.json
@@ -39,13 +39,12 @@
   },
   "dependencies": {
     "@auto-it/core": "link:../../packages/core",
+    "@auto-it/package-json-utils": "link:../../packages/package-json-utils",
     "await-to-js": "^2.1.1",
     "env-ci": "^5.0.1",
     "fp-ts": "^2.5.3",
     "get-monorepo-packages": "^1.1.0",
     "io-ts": "^2.1.2",
-    "parse-author": "^2.0.0",
-    "parse-github-url": "1.0.2",
     "registry-url": "^5.1.0",
     "semver": "^7.0.0",
     "tslib": "2.0.3",
@@ -56,7 +55,6 @@
   "devDependencies": {
     "@octokit/rest": "^18.0.0",
     "@types/env-ci": "^3.1.0",
-    "@types/parse-github-url": "1.0.0",
     "@types/semver": "^7.1.0",
     "@types/url-join": "^4.0.0",
     "@types/user-home": "^2.0.0"
