+++ tsconfig.dev.json
@@ -76,6 +76,9 @@
     },
     {
       "path": "plugins/microsoft-teams"
+    },
+    {
+      "path": "packages/package-json-utils"
     }
   ]
 }
